+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * ScenarioDbTable._set_df_column_types added support for Time type for time-of-day columns
 * DataManager.get_parameter_value added support for Time type for time-of-day parameters
+* Optimization progress tracking for lexicographical optimization in Core01CpoOptimizationEngine and Core02CpoOptimizationEngine
 ### Fixed
 * Pandas to_excel FutureWarning fix
+* Bugfix in ScenarioDbTable.set_df_column_types() on Time conversion
 
 ## [0.5.8.0]- 2026-01-23
 ### Fixed